Feature/7 develop rewards page (#25)
* feat: 홈 화면 구현

* feat: 카테고리 화면 구현

* feat: 상품 디테일화면 구현

* refactor: 상품 서비스분류 및 이미지 파싱 수정

* feat: 상품 리뷰페이지 구현

* feat: 리뷰사진 api 구현

* refactor: products 도메인 예외 처리

* refactor: products option_value 수정

* fix: products 관련 api 변경 및 요청 사항 수정

* refactor : /api/v1/products/category API 수정

- /api/v1/products/category?categoryId={id} => /api/v1/products/categories/{category_id}

* remove: SecurityConfig 제거

* feat: reward 조회 API 구현(현재,판매,구매)

* fix: reward_type property 수정

* feat: 리워드 내역 조회 기능 구현

* fix: prduct관련 예외처리 및 product_code 수정, 우선순위 조건 수정

* chore: jsonwebtoken 의존성 추가

* feat: jwt 환경 구축

* chore: openfeign 의존성 추가

* feat: 카카오 회원가입 / 로그인 기능 구현

* feat: refresh 토큰 기능 구현

* rename: 패키지명 oauth2 -> auth로 변경

- refresh 토큰 기능이 추가되면서 더 넓은 의미인 auth로 패키지 이름을 변경합니다.

* refactor: jwt 토큰 파싱에 대한 에러 핸들링 통합

- 프론트에서 jwt 토큰 파싱 중 SignatureException나 ExpiredJwtException와 같은 구체적인 에러를 알 필요는 없고, 유효하지 않는 토큰이라는 것을 알면 된다고 생각해 에러를 통합합니다.

* rename: 컨트롤러명 OAuth2Controller -> AuthController로 변경

- 패키지명과 함께 변경했어야 하는데 빠뜨려서 지금 반영합니다.

* refactor: errorCode를 HttpStatus의 값을 사용하도록 변경

* fix: swagger 주소는 인가 과정을 거치지 않도록 수정

* fix: 필터에서의 인증 객체 저장 로직 변경

- token 검증 후 인증 객체(사용자 정보를 담은)를 저장할 때 사용자 정보를 가져오기 위한 DB 접근 코드를 제거하고 token의 payload 정보만 사용하도록 변경합니다.
- [참고자료] https://velog.io/@cham/JWT-JWT-%EA%B3%A0%EC%B0%B0-SecurityJWT

* comment: 프론트와 의논이 필요한 작업에 대한 주석 추가

* refactor: 컨트롤러 반환 값 변경

* feat: 로그아웃 기능 구현

* feat: 회원 탈퇴 기능 구현

- 회원이 탈퇴 후 재가입을 했을 때 기존 DB의 데이터를 사용하려고 했습니다. 하지만, 탈퇴한 기간 동안 닉네임이나 프로필 사진 등 데이터가 변할 가능성이 있어 기존 재가입 로직은 제거합니다.

* refactor: UserDto 팩토리 메서드 인자 변경

* refactor: api 메서드 방식, url 주소 변경

* fix: 로그아웃된 사용자는 탈퇴가 되지 않도록 로직 추가

* feat: 회원 프로필 조회 기능 구현

* feat: 자신이 생성한 홍보 링크 리스트 조회 기능 구현

* feat: 자신의 주문 내역 전체 조회 API 구현

* refactor: 프로필 관련 DTO, 변수명 수정

* feat: 자신의 주문 내역 상세 조회 API 구현

* feat: 배송지 생성 API 구현

* fix: 프로필 관련 쿼리메서드 최적화

- join 쿼리가 나가지 않도록 엔티티로 연관관계 조회
- 프로필 관련 데이터 조회 전 해당 회원이 활성화된 회원인지 확인

* refactor: 프로필 관련 서비스들을 MyService 클래스로 이동

* feat: 배송지 리스트 조회 API 구현

* feat: 배송지 삭제 API 구현

* feat: 기본 배송지 조회 API 구현

* fix: 배송지 추가 / 삭제 관련 로직 수정

- 배송지 추가 시, 첫 배송지 추가라면 해당 배송지를 기본 배송지로 설정되도록 수정
- 배송지 삭제 시, 삭제할 배송지가 기본 배송지이고 다른 배송지가 있다면 가장 최근에 등록한 배송지를 기본 배송지로 설정되도록 수정

* feat: 기본 배송지 변경 API 구현

* fix: 쿼리메서드 , 매서드명, url 수정

* refactor : 현재 리워드 조회 api 레포지토리 변경

- 한번만 query 하더라도 변경

* feat: 리뷰 생성,수정,삭제 API 및 이미지 S3 업로드 구현

* feat: 인출하기 & 인출 내역 API 구현

* refactor: transaction 어노테이션 추가

* fix: service로직 transaction적용 및 user 엔티티 수정

* fix: log 추가 및 pageable 코드 수정

* fix: 마이페이지 api 변경 및 요청 사항 반영

* feat: 회원 배송지 수정 API 구현

* fix: 리뷰 사진 pageable 수정 및 S3의존성 수정

* refactor: Address 생성 DTO 수정, 주소 삭제 api 수정 및 ignore 추가

Address 생성 DTO
- valid : 정규화
   - 전화번호 : xxx-xxxx-xxxx 또는 xxx-xxx-xxxx만 받음
- delete 메소드에 해당 유저의 주소가 맞는 지 확인 하는 로직 추가
- ignore : application-payment.yml

* refactor : 엔티티 변경 및 dto 수정

reward
- reward_withdrawal_account 엔티티 관계 x -> reward_weithdrawal 엔티티 관계 OneToOne
- dto :
   - RewardDto
     - RewardType 속성 추가
     - RewardService mapRewardToDto 수정

reward_withdrawal
- withdrawal_status 속성 추가
  -> 관련 dto에 해당 속성 추가 withdrawalHistoryDto

reward_account
- reward 관계 삭제

product
- rate 관련 속성 int -> float
- int 타입 -> Integer
- DTO
  - ProductCategoryDto
    - product id -> productCode
 - ProductOptionDTO
   - value : List<String> -> String : 굳이 리스트로 한번 묶을 필요가 없어서 지움

* fix : withdrawalservice, ProductReviewRepository,  ProductsController 메소드 수정

- productsController 의 getReviewPhotos(포토 리뷰 가져오기) page : 1 -> 0 로 초기화
- withdrawalService
   - "인출 가능한 최소 리워드는 5000 이상이어야 합니다.":
       - currentRewardAmount -> withdrawalAmount : 현재 가지고 있는 돈을 확인 하는 것이 아니라 인출하려는 금액을 비교해야 됨
   - Reward 생성 시, 받은 리워드 금액(ReceivedRewardAmount, 마이너스로 저장, 인출로 인한 금액 차감) 추가
  - 기존 reward가 가지고 있는 리워드 계좌를 저장하면 안됨 및 인출 엔티티 추가됨 : newRewardWithdrawal 저장

* feat: 리뷰 업로드(s3) 수정

* refactor : ignore 추가

- application-payment 추가

* refactor: product 리스트 속성에 wish 추가

* refactor : api 수정, dto 수정 & fix: reward 반환 수정

refactor : product-category api -> categories api로 변경
refactor : productDetailDto, ProductDto : reward -> rewarded
fix: reward : product.sale_price * reward_rate(범위 0 < x <=1)

* fix: ReviewPhoto 전송 갯수 수정

* fix: product_detail dto수정

* fix: 토큰을 dto가 아닌 cookie로 담아서 반환되도록 수정

* fix: 회원가입 시 프로필 사진이 기본 이미지가 되도록 수정

* feat: S3 파일 업로드 / 삭제 기능 구현

* feat: 기본 프로필 사진으로 변경 API 구현

* feat: 원하는 프로필 사진으로 변경 API 구현

* fix: JWT 에러 환경, 로그 수정

* refactor: products, rewards 도메인 예외처리

* fix: photo-reviews API 반환값 수정

* refactor: /api/v1/products/{product_code} dto에 detailImagePaths 추가

* refactor: category api 이전

* refactor: 배송지 관련 변수명, 클래스명

* refactor: 불필요한 fetch join 제거

* feat: review scheduler 구현

* refactor: JPA 페이징 처리 최적화

- 추가적인 count() 쿼리가 실행되지 않도록 페이징 타입을 Page에서 List로 변경

* refactor: LocalDateTime Java17 버전으로 수정

* refactor: 에러 처리 및 예외 처리 추가

- BaseException (의도된 예외 및 에러)
- 그 이외의 에러 (의도치 않은 에러)
- try ~ catch 로 감쌈

* fix: controller 내 try ~ catch 제거

- controller 내 try ~ catch 넣기 보단 service 내에서 넣는 게 맞다고 봐 제거 했습니다

* refactor: 리워드 및 인출 관련 코드 수정

- reward, widthdrawal 엔티티 -> builder
- star : double -> float
- StarRate : double -> float
- 예외 처리 추가
- repository 일부 메소드 Optional 추가

* feat: 3개월 단위로 리뷰(deletedAt !=null) 삭제

* refactor: 기능 추가

* Refactor: /api/my 관련 API 예외 처리 추가

try ~ catch로 묶으고 BaseException의 경우 BaseException 에러 반환 그 이외의 예러도 BaseException의 반환

* Refactor: Access 및 Refresh 토큰 값 application으로 저장

* feat: 금융결제원 Api 사용자인증 및 token 발급 구현

* fix: OpenBankingDto 생성자 추가 및 이름변경

* 임시

- 삭제할 커밋

* feat: 금융결제원 계좌실명조회 API 구현 및 테스트코드 작성

* refactor: 배송지 관련 코드 정리

* test: 배송지 관련 API 테스트 코드 작성

* test: 인증 관련 API 테스트 코드 작성

- 테스트 코드 작성하면, MyService에 있는 인증 관련 메서드 AuthService로 이동시켰습니다.

* 임시

- 삭제할 커밋

* test: 프로필 관련 API 테스트 코드 작성

* fix: cookie의 secure 속성 비할성화

* fix: 토큰 관련 API 요청과 응답 방식 수정

- 토큰 재발급 시 refresh 토큰을 헤더로 받도록
- 토큰 생성 후 응답 시 access 토큰은 바디에 refresh 토큰은 쿠키에 담아 반환되도록

* feat: 계좌 정보 가져오기 API 구현

* fix: product entity 수정

* refactor: 프로필의 rewardSum 조회 시 쿼리메서드 사용

* feat: rewards 도메인 테스트 코드 작성

* refactor: withdrawalHistoryDto -> withdrawalDto

* rename: 예외 관련 파일 적절한 위치로 이동

* fix: 잘못된 위치에 작성된 코드 제거

* feat: 재가입 로직 구현 및 테스트 코드 작성

* fix: 불필요한 쿼리문 조건 제거

* refactor: 자잘한 테스트 코드 수정

* test: contextLoads 빌드 실패 해결

- test 환경파일이 없어 전체 테스트가 실패해 test 환경의 yml 추가 및 프로파일 적용합니다.

* feat: 공동구매자DTO 추가

* chore: swagger security 의존성 추가

- 해당 의존성을 추가하면, @AuthenticationPrincipal CustomUserDetails userDetails 파라미터를 자동으로 swagger API 문서에 보이지 않도록 해준다.

* fix: 불필요한 코드 삭제 및 수정

* feat: 연관관계 One 방향에 lazy 조회 속성 추가

* refactor: page 0으로 초기화

* refactor: 내 주문 상세에 썸네일 이미지 추가

* bug: page 0으로 초기화

* refactor: merge를 위한 product 관련 app 제거

- product 관련 controller, service, dto 제거

* refactor: userId mock 값 userDetails 적용

* refacotr: category api userId 모의 값 userDetails 로 변경

* refactor: user 모의 객체 실제 user 객체로 변경

* refactor: Component(Controller, Service, component) requiredArgsConstructor 적용

* refactor: reward Api 관련 component(Controller, Service)에 RequiredArgConstructor 적용

* refactor: reward 관련 api에 userDetails 적용

* bug: 배송지 삭제시 현재 배송지 삭제 못하게 하기

* refactor: 프로필 이미지 변경 api 반환값 dto 추가 및 반환값 변경

* refactor: 프로필 기본 이미지 변경 및 user 엔티티 profile_path 길이 5000으로 변경

* refactor: my 관련 api에 반환값 타입 지정

* feat: 리워드 계좌 조회 API 및 수정 API 개발

* fix: renwal API 요청시 리프레쉬 토큰 쿠키 안 받아와지는 버그 해결 및 로그아웃 API 할시 리프레쉬 토큰 쿠키 제거 되도록 수정

* fix: 기본 주소지 조회 url 변경

* fix: 금융결제원 관련 api 수정

- 나중에 한번 봐야 될듯

* fix(test): 테스트 코드 수정

* build: 필요없는(중복되는) 의존성 제거

- spring-cloud-starte-aws 의존성 제거

---------

+++ src/main/java/com/freeder/buclserver/domain/productreview/dto/ReviewDTO.java
@@ -12,7 +12,7 @@ public class ReviewDTO {
 	private String profilePath;
 	private String nickname;
 	private LocalDateTime createdAt;
-	private double starRate;
+	private float starRate;
 	private String selectedOption;
 	private List<String> reviewImages;
 	private String reviewText;
